Fix for incorrect call of the function  _merge_buffer_entries()
In postprocess.py the function _merge_buffer_entries() called using "self." notation, but it is not member of any class, so i assume it should be called as _merge_buffer_entries(self, ...). Similar to other functions in this module. Otherwise it leads to AttributeError: 'ModelHistory' object has no attribute '_merge_buffer_entries'